+++ src/test/java/nom/bdezonia/zorbage/algorithm/TestIsPartitioned.java
@@ -53,13 +53,13 @@ public class TestIsPartitioned {
 		IndexedDataSource<?, SignedInt8Member> a = ArrayStorage.allocateBytes(
 				new byte[] {1,2,3,4,5,6,7,8});
 		
-		Condition<SignedInt8Member> cond1 = new EqualConstant<SignedInt8Group, SignedInt8Member>(G.INT8, new SignedInt8Member((byte)4));
+		Condition<SignedInt8Member> cond1 = new EqualConstant<SignedInt8Group, SignedInt8Member>(G.INT8, new SignedInt8Member(4));
 		assertFalse(IsPartitioned.compute(G.INT8, cond1, a));
 
-		Condition<SignedInt8Member> cond2 = new LessThanConstant<SignedInt8Group, SignedInt8Member>(G.INT8, new SignedInt8Member((byte)4));
+		Condition<SignedInt8Member> cond2 = new LessThanConstant<SignedInt8Group, SignedInt8Member>(G.INT8, new SignedInt8Member(4));
 		assertTrue(IsPartitioned.compute(G.INT8, cond2, a));
 
-		Condition<SignedInt8Member> cond3 = new GreaterThanConstant<SignedInt8Group, SignedInt8Member>(G.INT8, new SignedInt8Member((byte)4));
+		Condition<SignedInt8Member> cond3 = new GreaterThanConstant<SignedInt8Group, SignedInt8Member>(G.INT8, new SignedInt8Member(4));
 		assertTrue(IsPartitioned.compute(G.INT8, cond3, a));
 	}
 }
